Add localStorage
Update index.js to implement the storage of the data into localStorage and the render of the data saved when the page is loaded.

Update style.css to fix a visual issue.

Create todolist.js module, to separate the class todoList out of the index.js.
