+++ src/style.css
@@ -77,7 +77,7 @@ h1 {
 }
 
 #list-input-left {
-  width: 90%;
+  width: 80%;
 }
 
 #list-input-left input[type='text'] {
